backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.12.3
bleach==6.2.0
blivet==3.11.0
blivet-gui==2.6.0
boto3==1.36.22
botocore==1.36.22
Brlapi==0.8.5
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnf==4.22.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.15.4
fqdn==1.5.1
fros==1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.9.0
idna==3.7
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
iso639==0.1.4
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
langtable==0.0.68
libcomps==0.1.21
libdnf==0.73.4
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mod_wsgi==5.0.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nftables==0.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.3
olefile==0.47
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
platformdirs==4.3.6
ply==3.11
productmd==1.43
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pyasn1==0.4.8
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyenchant==3.2.2
Pygments==2.19.1
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.4.0
python-json-logger==3.2.1
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
python-slugify==8.0.4
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.0
rpm==4.20.0
rsa==4.9.1
s3transfer==0.11.2
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
Send2Trash==1.8.3
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.8.2
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
systemd-python==235
Tempita==0.5.2
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.21.1
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
xkbregistry==0.3
